+++ src/token/TokenUnpauseTransaction.js
@@ -1,0 +1,148 @@
+import TokenId from "./TokenId.js";
+import Transaction, {
+    TRANSACTION_REGISTRY,
+} from "../transaction/Transaction.js";
+
+/**
+ * @namespace proto
+ * @typedef {import("@hashgraph/proto").ITransaction} proto.ITransaction
+ * @typedef {import("@hashgraph/proto").ISignedTransaction} proto.ISignedTransaction
+ * @typedef {import("@hashgraph/proto").TransactionBody} proto.TransactionBody
+ * @typedef {import("@hashgraph/proto").ITransactionBody} proto.ITransactionBody
+ * @typedef {import("@hashgraph/proto").ITransactionResponse} proto.ITransactionResponse
+ * @typedef {import("@hashgraph/proto").ITokenUnpauseTransactionBody} proto.ITokenUnpauseTransactionBody
+ * @typedef {import("@hashgraph/proto").ITokenID} proto.ITokenID
+ */
+
+/**
+ * @typedef {import("../channel/Channel.js").default} Channel
+ * @typedef {import("../client/Client.js").default<*, *>} Client
+ * @typedef {import("../transaction/TransactionId.js").default} TransactionId
+ * @typedef {import("../account/AccountId.js").default} AccountId
+ */
+
+/**
+ * Unpause a new Hedera™ crypto-currency token.
+ */
+export default class TokenUnpauseTransaction extends Transaction {
+    /**
+     * @param {object} [props]
+     * @param {TokenId | string} [props.tokenId]
+     */
+    constructor(props = {}) {
+        super();
+
+        /**
+         * @private
+         * @type {?TokenId}
+         */
+        this._tokenId = null;
+
+        if (props.tokenId != null) {
+            this.setTokenId(props.tokenId);
+        }
+    }
+
+    /**
+     * @internal
+     * @param {proto.ITransaction[]} transactions
+     * @param {proto.ISignedTransaction[]} signedTransactions
+     * @param {TransactionId[]} transactionIds
+     * @param {AccountId[]} nodeIds
+     * @param {proto.ITransactionBody[]} bodies
+     * @returns {TokenUnpauseTransaction}
+     */
+    static _fromProtobuf(
+        transactions,
+        signedTransactions,
+        transactionIds,
+        nodeIds,
+        bodies
+    ) {
+        const body = bodies[0];
+        const unpauseToken = /** @type {proto.ITokenUnpauseTransactionBody} */ (
+            body.tokenUnpause
+        );
+
+        return Transaction._fromProtobufTransactions(
+            new TokenUnpauseTransaction({
+                tokenId:
+                    unpauseToken.token != null
+                        ? TokenId._fromProtobuf(unpauseToken.token)
+                        : undefined,
+            }),
+            transactions,
+            signedTransactions,
+            transactionIds,
+            nodeIds,
+            bodies
+        );
+    }
+
+    /**
+     * @returns {?TokenId}
+     */
+    get tokenId() {
+        return this._tokenId;
+    }
+
+    /**
+     * @param {TokenId | string} tokenId
+     * @returns {this}
+     */
+    setTokenId(tokenId) {
+        this._requireNotFrozen();
+        this._tokenId =
+            typeof tokenId === "string"
+                ? TokenId.fromString(tokenId)
+                : tokenId.clone();
+
+        return this;
+    }
+
+    /**
+     * @param {Client} client
+     */
+    _validateChecksums(client) {
+        if (this._tokenId != null) {
+            this._tokenId.validateChecksum(client);
+        }
+    }
+
+    /**
+     * @override
+     * @internal
+     * @param {Channel} channel
+     * @param {proto.ITransaction} request
+     * @returns {Promise<proto.ITransactionResponse>}
+     */
+    _execute(channel, request) {
+        return channel.token.unpauseToken(request);
+    }
+
+    /**
+     * @override
+     * @protected
+     * @returns {NonNullable<proto.TransactionBody["data"]>}
+     */
+    _getTransactionDataCase() {
+        return "tokenUnpause";
+    }
+
+    /**
+     * @override
+     * @protected
+     * @returns {proto.ITokenUnpauseTransactionBody}
+     */
+    _makeTransactionData() {
+        return {
+            token: this._tokenId != null ? this._tokenId._toProtobuf() : null,
+        };
+    }
+}
+
+TRANSACTION_REGISTRY.set(
+    "tokenUnpause",
+    // eslint-disable-next-line @typescript-eslint/unbound-method
+    TokenUnpauseTransaction._fromProtobuf
+);